moved server functions to a seperate file
