+++ src/main.cpp
@@ -1,11 +1,11 @@
 #include <ESP8266WiFi.h>
 #include <DNSServer.h>
 #include <ESP8266WebServer.h>
+#include "webserver.h"
 
 const byte DNS_PORT = 53;
 IPAddress apIP(172, 217, 28, 1);
 DNSServer dnsServer;
-ESP8266WebServer webServer(80);
 
 #include <LittleFS.h>
 // const char* fsName = "LittleFS";
@@ -18,20 +18,6 @@ String responseHTML = ""
                       "<title>CaptivePortal</title></head><body>"
                       "<h1>Hello World!</h1><p>This is a captive portal example."
                       " All requests will be redirected here.</p></body></html>";
-
-bool handleCaptivePortal() { // send the right file to the client (if it exists)
-  const char* path = "/config.html";
-  String contentType = "text/html";            // Get the MIME type
-  if (LittleFS.exists(path)) {                            // If the file exists
-    File file = LittleFS.open(path, "r");                 // Open it
-    webServer.streamFile(file, contentType); // And send it to the client
-    Serial.println("file sent to client");
-    file.close();                                       // Then close the file again
-    return true;
-  }
-  Serial.println("\tFile Not Found");
-  return false;                                         // If the file doesn't exist, return false
-}
 
 void listDir(const char * dirname) {
   Serial.printf("Listing directory: %s\n", dirname);
@@ -66,22 +52,13 @@ void setup() {
     return;
   }
 
-  listDir("/");
+  // listDir("/");
 
-  webServer.on("/captive_portal", handleCaptivePortal);
-  // replay to all requests with same HTML
-  webServer.onNotFound([]() {
-    Serial.println("handling Not Found");
-    String message = "URI: ";
-    message += webServer.uri();
-    Serial.println(message);
-    handleCaptivePortal();
-    // webServer.send(200, "text/plain", "");
-  });
-  webServer.begin();
+  setupWebServer();
 }
 
 void loop() {
   dnsServer.processNextRequest();
-  webServer.handleClient();
+  // webServer.handleClient();
+  handlePortalClient();
 }
